refactor: bitcoin api config
